+++ src/engine/TeachingPlayground.ts
@@ -3,7 +3,7 @@ import { RealTimeCommunicationSystem } from '../systems/comms/RealTimeCommunicat
 import { DataManagementSystem } from '../systems/data/DataManagementSystem'
 import { EventManagementSystem } from '../systems/event/EventManagementSystem'
 import { RoomManagementSystem } from '../systems/room/RoomManagementSystem'
-import { Lecture, LectureReservation, ReservationAcademicPath, ReservationFilter } from '../interfaces/event.interface'
+import { AttendanceReport, AttendanceSnapshot, CaptureAttendanceSnapshotOptions, Lecture, LectureReservation, RecordAttendanceEventOptions, ReservationAcademicPath, ReservationFilter } from '../interfaces/event.interface'
 import { SystemError } from '../interfaces'
 import { User, TeacherProfile } from '../interfaces/user.interface'
 import { RoomFeatures } from '../interfaces/room.interface'
@@ -127,6 +127,23 @@ export default class TeachingPlayground {
 
   async cancelReservation(lectureId: string, reason?: string) {
     return this.eventSystem.cancelReservation(lectureId, this.requireOrganization(), reason)
+  }
+
+
+  async recordAttendanceEvent(options: Omit<RecordAttendanceEventOptions, 'organizationId'>) {
+    return this.eventSystem.recordAttendanceEvent({ ...options, organizationId: this.requireOrganization() })
+  }
+
+  async captureAttendanceSnapshot(options: Omit<CaptureAttendanceSnapshotOptions, 'organizationId'>): Promise<AttendanceSnapshot> {
+    return this.eventSystem.captureAttendanceSnapshot({ ...options, organizationId: this.requireOrganization() })
+  }
+
+  async finalizeAttendanceReport(reservationId: string): Promise<AttendanceReport> {
+    return this.eventSystem.finalizeAttendanceReport(this.requireOrganization(), reservationId)
+  }
+
+  async getAttendanceReport(reservationId: string): Promise<AttendanceReport | null> {
+    return this.eventSystem.getAttendanceReport(this.requireOrganization(), reservationId)
   }
 
   // Enhanced Event Management